+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 - Properly clear blacklisted IPs when updating blacklist
 - Update blacklist even when server is paused
+- Add config for blacklist update cooldown